fix(code-review): fix code review
- let empty json in the locales folder
- add decorators folder in storybook
- delete the <p></p> tag in app.tsx
