+++ .storybook/preview.tsx
@@ -1,24 +1,6 @@
 import type { Preview } from '@storybook/react';
 import { themes } from '@storybook/theming';
-import { I18nextProvider } from 'react-i18next';
-import i18n from '../src/i18n';
-import React, { Suspense, useEffect } from 'react';
-
-const withI18next = (Story, context) => {
-  const { locale } = context.globals;
-
-  useEffect(() => {
-    i18n.changeLanguage(locale);
-  }, [locale]);
-
-  return (
-    <Suspense fallback={<div>loading translations...</div>}>
-      <I18nextProvider i18n={i18n}>
-        <Story />
-      </I18nextProvider>
-    </Suspense>
-  );
-};
+import { withI18next } from './decorators/i18n';
 
 export const globalTypes = {
   locale: {
