+++ locales/en.json
@@ -1,5 +1,1 @@
-{
-
-      "Welcome to React": "Welcome to React Dude"
-
-  }
+{}
