+++ locales/pt.json
@@ -1,5 +1,1 @@
-{
-
-    "Welcome to React": "Bem-vindo ao React"
-
-}
+{}
