+++ src/App.tsx
@@ -1,13 +1,11 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import { useTranslation } from 'react-i18next';
-import './App.css'
+import { useState } from 'react';
+import reactLogo from './assets/react.svg';
+import viteLogo from '/vite.svg';
 
+import './App.css';
 
 function App() {
-  const [count, setCount] = useState(0)
-  const {t} = useTranslation();
+  const [count, setCount] = useState(0);
 
   return (
     <>
@@ -21,18 +19,15 @@ function App() {
       </div>
       <h1>Vite + React</h1>
       <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
+        <button onClick={() => setCount(count => count + 1)}>
           count is {count}
         </button>
         <p>
           Edit <code>src/App.tsx</code> and save to test HMR
         </p>
       </div>
-      <p className="read-the-docs">
-       {t('Welcome to React', {lng: 'en'})}
-      </p>
     </>
-  )
+  );
 }
 
-export default App
+export default App;
